Refacto + use snackbar hook in profile page
